[IMP] l10n_in_edi_ewaybill: remove dependency of l10n_in_edi
