+++ addons/l10n_in/data/neutralize.sql
@@ -1,0 +1,2 @@
+UPDATE res_company
+   SET l10n_in_edi_env = 'testing';
