+++ addons/l10n_in/models/company.py
@@ -12,6 +12,13 @@ class ResCompany(models.Model):
         ],
         string="HSN Code Digit",
     )
-    module_l10n_in_edi = fields.Boolean(string="Enable Indian Electronic Invoicing")
-    module_l10n_in_edi_ewaybill = fields.Boolean(string="Enable Indian Electronic WayBill")
-    module_l10n_in_reports_gstr = fields.Boolean(string="Indian GST Service")
+    l10n_in_edi_env = fields.Selection(
+        selection=[
+            ("testing", "Testing"),
+            ("production", "Production"),
+        ],
+        string="Indian EDI Environment",
+        help="Enable the use of production credentials",
+        default="testing",
+        groups="base.group_system",
+    )
