+++ addons/l10n_in_edi/__manifest__.py
@@ -6,9 +6,7 @@
     'countries': ['in'],
     "category": "Accounting/Localizations/EDI",
     "depends": [
-        "account_edi",
         "l10n_in",
-        "iap",
     ],
     "description": """
 Indian - E-invoicing
