+++ addons/l10n_in_edi/data/neutralize.sql
@@ -1,7 +1,6 @@
 -- disable l10n_in_edi integration
 UPDATE res_company
-   SET l10n_in_edi_env = 'testing',
-       l10n_in_edi_username = NULL,
+   SET l10n_in_edi_username = NULL,
        l10n_in_edi_password = NULL,
        l10n_in_edi_token = NULL,
        l10n_in_edi_token_validity = NULL;
