+++ addons/l10n_in_edi/i18n/l10n_in_edi.pot
@@ -127,7 +127,7 @@ msgid "Duplicate"
 msgstr ""
 
 #. module: l10n_in_edi
-#: model:ir.model.fields,field_description:l10n_in_edi.field_res_company__l10n_in_edi_env
+#: model:ir.model.fields,field_description:l10n_in_edi.field_res_company__l10n_in_edi_production_env
 msgid "E-invoice (IN) Is production OSE environment"
 msgstr ""
 
@@ -164,8 +164,8 @@ msgid "EDI format"
 msgstr ""
 
 #. module: l10n_in_edi
-#: model:ir.model.fields,help:l10n_in_edi.field_res_company__l10n_in_edi_env
-#: model:ir.model.fields,help:l10n_in_edi.field_res_config_settings__l10n_in_edi_env
+#: model:ir.model.fields,help:l10n_in_edi.field_res_company__l10n_in_edi_production_env
+#: model:ir.model.fields,help:l10n_in_edi.field_res_config_settings__l10n_in_edi_production_env
 msgid "Enable the use of production credentials"
 msgstr ""
 
@@ -195,8 +195,8 @@ msgid ""
 msgstr ""
 
 #. module: l10n_in_edi
-#: model:ir.model.fields,field_description:l10n_in_edi.field_res_config_settings__l10n_in_edi_env
-msgid "Indian EDI Environment"
+#: model:ir.model.fields,field_description:l10n_in_edi.field_res_config_settings__l10n_in_edi_production_env
+msgid "Indian EDI Testing Environment"
 msgstr ""
 
 #. module: l10n_in_edi
