+++ addons/l10n_in_edi/models/res_company.py
@@ -11,16 +11,6 @@ class ResCompany(models.Model):
     l10n_in_edi_password = fields.Char("E-invoice (IN) Password", groups="base.group_system")
     l10n_in_edi_token = fields.Char("E-invoice (IN) Token", groups="base.group_system")
     l10n_in_edi_token_validity = fields.Datetime("E-invoice (IN) Valid Until", groups="base.group_system")
-    l10n_in_edi_env = fields.Selection(
-        selection=[
-            ("testing", "Testing"),
-            ("production", "Production"),
-        ],
-        string="E-invoice (IN) OSE environment",
-        help="Enable the use of production credentials",
-        default="testing",
-        groups="base.group_system",
-    )
 
     def _l10n_in_edi_token_is_valid(self):
         self.ensure_one()
