+++ addons/l10n_in_edi/models/res_config_settings.py
@@ -2,7 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import models, fields, _
-from odoo.exceptions import UserError, RedirectWarning
+from odoo.exceptions import UserError
 
 
 class ResConfigSettings(models.TransientModel):
@@ -10,22 +10,6 @@ class ResConfigSettings(models.TransientModel):
 
     l10n_in_edi_username = fields.Char("Indian EDI username", related="company_id.l10n_in_edi_username", readonly=False)
     l10n_in_edi_password = fields.Char("Indian EDI password", related="company_id.l10n_in_edi_password", readonly=False)
-    l10n_in_edi_env = fields.Selection(
-        string="Indian EDI Environment",
-        related="company_id.l10n_in_edi_env",
-        readonly=False
-    )
-
-    def l10n_in_check_gst_number(self):
-        if not self.company_id.vat:
-            action = {
-                    "view_mode": "form",
-                    "res_model": "res.company",
-                    "type": "ir.actions.act_window",
-                    "res_id" : self.company_id.id,
-                    "views": [[self.env.ref("base.view_company_form").id, "form"]],
-            }
-            raise RedirectWarning(_("Please enter a GST number in company."), action, _('Go to Company'))
 
     def l10n_in_edi_test(self):
         self.l10n_in_check_gst_number()
@@ -41,12 +25,3 @@ class ResConfigSettings(models.TransientModel):
                   'message': _("API credentials validated successfully"),
               }
           }
-
-    def l10n_in_edi_buy_iap(self):
-        if not self.l10n_in_edi_env == 'production':
-            raise UserError(_("You must enable production environment to buy credits"))
-        return {
-            'type': 'ir.actions.act_url',
-            'url': self.env["iap.account"].get_credits_url(service_name="l10n_in_edi", base_url=''),
-            'target': '_new'
-        }
