+++ addons/l10n_in_edi_ewaybill/views/res_config_settings_views.xml
@@ -5,7 +5,7 @@
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="account.res_config_settings_view_form"/>
         <field name="arch" type="xml">
-            <xpath expr="//setting[@id='l10n_in_eway_iap']" position="inside">
+            <xpath expr="//setting[@id='l10n_in_edi_ewaybill_setting']" position="inside">
                 <div class="content-group">
                     <div class="mt16 row">
                         <label for="l10n_in_edi_ewaybill_username" string="Username" class="col-3 col-lg-3 o_light_label"/>
@@ -15,15 +15,8 @@
                         <label for="l10n_in_edi_ewaybill_password" string="Password" class="col-3 col-lg-3 o_light_label"/>
                         <field name="l10n_in_edi_ewaybill_password" password="True" nolabel="1"/>
                     </div>
-                    <div class="row">
-                        <label for="l10n_in_edi_env" string="Environment" class="col-3 col-lg-3 o_light_label"/>
-                        <field name="l10n_in_edi_env" nolabel="1" widget="radio" options="{'horizontal':true}"/>
-                    </div>
                     <div class='mt8'>
                         <button name="l10n_in_edi_ewaybill_test" icon="oi-arrow-right" type="object" string="Verify Username and Password" class="btn-link"/>
-                    </div>
-                    <div class='mt8'>
-                        <button name="l10n_in_edi_buy_iap" title="Costs 1 credit per transaction. Free 200 credits will be available for the first time." icon="oi-arrow-right" type="object" string="Buy credits" class="btn-link"/>
                     </div>
                 </div>
             </xpath>
